This merge introduce the message routing protocol.

Main changes are:
 - When possible RemoteObjectFactory is in charge of creating the URL
   Clients gives a name for the object and the ROF returns an URL.
   Some methods taking an URL in NodeFactory and PAActiveObject have been deprecated
 - It is now possible to register a remote object protocol as a plugin
 - Few helper classes like Sleeper, SweetCountDownLatch have been added
 - Functional test can now be run using several protocols: junit.rmi
   junit.http etc.

Message routing should be functional. Its protocol ID is "pamr".
Lookups are not currently supported.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@11270 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Tests/functionalTests/component/collectiveitf/multicast/ServerImpl.java
@@ -78,7 +78,7 @@ public class ServerImpl implements ServerTestItf, Identifiable {
      * @see functionalTests.component.collectiveitf.multicast.ServerTestItf#testOneToOne(java.lang.MyObject)
      */
     public WrappedInteger testOneToOne_Param(WrappedInteger a) {
-        Assert.assertEquals(a.getIntValue(), id);
+        Assert.assertEquals(a.getIntValue().intValue(), id);
         return new WrappedInteger(id);
     }
